Current progress on the text finder, as of January
This is the current progress on the portion of the translator that performs document layout analysis and extracts the text regions in the page along with metadata describing their location (and perhaps type, ie whether they are from a speech bubble, or free-form text) so that they can be OCR'd.

Since the OCR required is to be performed by Tesseract, and translation is to be performed by an external service, like google translate, this forms the meat of the translator's functionality. It's possible that I may (at least temporarily) release a version where this job can bypassed and performed by a human, if real-world tests prove it to be unreliable.

Current implementation is in Turring 4.1.1, but will be translated into C when more feature-complete. The choice of Turing is currently for the purposes of having a safer language to prototype in which is easier to debug. An added side benefit is the poor performance of the language, especially when drawing a